requirements.txt
Django==4.2
djangorestframework==3.14.0
requests==2.29.0
djoser==2.2.0